adds view profile page, updates routes
