+++ hackathon-fe/src/App.js
@@ -8,6 +8,8 @@ import ProtectedPage from "./components/ProtectedPage/ProtectedPage";
 import LandingPage from "./pages/Landing/Landing";
 import SignUpPage from "./pages/SignUp/SignUp";
 import LoginPage from "./pages/Login/Login";
+import UpdateProfilePage from "./pages/Profile/UpdateProfile";
+import EditInterestsPage from './pages/Profile/EditInterests';
 import ProfilePage from "./pages/Profile/Profile";
 import FriendsPage from "./pages/FriendsPage/FriendsPage";
 import FriendPage from "./pages/FriendPage/FriendPage";
@@ -32,6 +34,22 @@ function App() {
             element={
               <ProtectedPage>
                 <ProfilePage />
+              </ProtectedPage>
+            }
+          />
+          <Route
+            path="/profile/edit"
+            element={
+              <ProtectedPage>
+                <UpdateProfilePage />
+              </ProtectedPage>
+            }
+          />
+          <Route
+            path="/profile/edit/interests"
+            element={
+              <ProtectedPage>
+                <EditInterestsPage />
               </ProtectedPage>
             }
           />
